Fix crashes when a Lua chunk does not contain a function of the same name.
When a Lua chank has a name, after the compilation an extended type of the
variable with the chunk's name is stored in the CtrlrLuaMethod object.
Errors are defined as additional types. This complies to the fact that a
Lua chunk can contain everything, not only a single functions.

The function isValid() changed its semantics from “running fine, so far”
to “no probable error has been detected, so far“. In this sense a named Lua
chunk that does not define a variable with the chunk's name is considered to
probably contain an error. The function isValid() is mainly used to inform
the Ctrlr user/programmer that a certain chunk is problematic in some way,
which meens colouring its name red if isValid() returns false.

In order to invalid calls to the Lua variable a function isCallable() has
been introduced. It returns true iff the Lua chunk defines a function with
the chunk's name that compiles fine and did not raise any Lua error so far.
So, isCallable() covers most of the cases where isValid() had been called
until this patch.

In order to avoid unnecessary confusion “json4lua“ has been renamed to “json”
and “debugger.lua” has been renamed to “debugger”.