Fix deletion of the MIDI snapshot thread.
A thread should be running either as detached thread or as joinable thread.
JUCE blurres this nomenclature with their own words.

In general creating and deleting a system thread is expensive.
So it should be deleted when it is save (i.e. not in realtime functions).

This patch solves several issues with the thread:
- the thread should exit if it gets signalled by setting a special
  variable. This variable is handled by JUCE.
- The main program should signal the thread to end its action. This is done
  by stopThread() by setting the status variable and waking up the thread.
- When the thread ends before the main program allows it to end,
  it must wait indefinitely. Otherwise an undefined state with a race
  condition is created, which may cause additional trouble.